feat: adicionando utils
